+++ metacells/pipeline/feature.py
@@ -35,8 +35,8 @@ def extract_feature_data(
     min_gene_top3: Optional[int] = pr.feature_min_gene_top3,
     forced_gene_names: Optional[Collection[str]] = None,
     forced_gene_patterns: Optional[Collection[Union[str, Pattern]]] = None,
-    forbidden_gene_names: Optional[Collection[str]] = None,
-    forbidden_gene_patterns: Optional[Collection[Union[str, Pattern]]] = None,
+    lateral_gene_names: Optional[Collection[str]] = None,
+    lateral_gene_patterns: Optional[Collection[Union[str, Pattern]]] = None,
     random_seed: int = 0,
     top_level: bool = True,
 ) -> Optional[AnnData]:
@@ -70,8 +70,8 @@ def extract_feature_data(
             A boolean mask of genes with "high" normalized variance, relative to other genes with a
             similar expression level.
 
-        ``forbidden_gene``
-            A boolean mask of genes which are forbidden from being chosen as "feature" genes based
+        ``lateral_gene``
+            A boolean mask of genes which are lateral from being chosen as "feature" genes based
             on their name.
 
         ``forced_gene``
@@ -89,7 +89,7 @@ def extract_feature_data(
        {downsample_max_cell_quantile}) and the ``random_seed`` (default: {random_seed}).
 
     2. Invoke :py:func:`metacells.tools.named.find_named_genes` to force genes as being used as
-       features based on their name, using ``forced_gene_names`` and ``forbidden_gene_patterns``.
+       features based on their name, using ``forced_gene_names`` and ``lateral_gene_patterns``.
 
     3. Invoke :py:func:`metacells.tools.high.find_high_total_genes` to select high-expression
        feature genes (based on the downsampled data), using ``min_gene_total``.
@@ -99,9 +99,9 @@ def extract_feature_data(
        ``min_gene_relative_variance``.
 
     5. Invoke :py:func:`metacells.tools.named.find_named_genes` to forbid genes from being used as
-       feature genes, based on their name, using the ``forbidden_gene_names`` (default:
-       {forbidden_gene_names}) and ``forbidden_gene_patterns`` (default: {forbidden_gene_patterns}).
-       This is stored in an intermediate per-variable (gene) ``forbidden_genes`` boolean mask.
+       feature genes, based on their name, using the ``lateral_gene_names`` (default:
+       {lateral_gene_names}) and ``lateral_gene_patterns`` (default: {lateral_gene_patterns}).
+       This is stored in an intermediate per-variable (gene) ``lateral_genes`` boolean mask.
 
     6. Invoke :py:func:`metacells.tools.filter.filter_data` to slice just the selected
        "feature" genes using the ``name`` (default: {name}).
@@ -135,9 +135,9 @@ def extract_feature_data(
             adata, "downsampled", min_gene_relative_variance=min_gene_relative_variance
         )
 
-    if forbidden_gene_names is not None or forbidden_gene_patterns is not None:
-        var_masks.append("~forbidden_gene")
-        tl.find_named_genes(adata, to="forbidden_gene", names=forbidden_gene_names, patterns=forbidden_gene_patterns)
+    if lateral_gene_names is not None or lateral_gene_patterns is not None:
+        var_masks.append("~lateral_gene")
+        tl.find_named_genes(adata, to="lateral_gene", names=lateral_gene_names, patterns=lateral_gene_patterns)
 
     results = tl.filter_data(
         adata, name=name, top_level=top_level, track_var="full_gene_index", mask_var="feature_gene", var_masks=var_masks