+++ metacells/pipeline/related_genes.py
@@ -42,8 +42,8 @@ def relate_genes(
     min_gene_relative_variance: float = pr.related_min_gene_relative_variance,
     min_gene_total: int = pr.related_min_gene_total,
     min_gene_top3: int = pr.related_min_gene_top3,
-    forbidden_gene_names: Optional[Collection[str]] = None,
-    forbidden_gene_patterns: Optional[Collection[Union[str, Pattern]]] = None,
+    lateral_gene_names: Optional[Collection[str]] = None,
+    lateral_gene_patterns: Optional[Collection[Union[str, Pattern]]] = None,
     genes_similarity_method: str = pr.related_genes_similarity_method,
     genes_cluster_method: str = pr.related_genes_cluster_method,
     min_genes_of_modules: int = pr.related_min_genes_of_modules,
@@ -105,8 +105,8 @@ def relate_genes(
         min_gene_relative_variance=min_gene_relative_variance,
         min_gene_total=min_gene_total,
         min_gene_top3=min_gene_top3,
-        forbidden_gene_names=forbidden_gene_names,
-        forbidden_gene_patterns=forbidden_gene_patterns,
+        lateral_gene_names=lateral_gene_names,
+        lateral_gene_patterns=lateral_gene_patterns,
         random_seed=random_seed,
     )
     assert fdata is not None