+++ metacells/tools/quality.py
@@ -557,8 +557,8 @@ def compute_metacells_projection_correlation(
         "&atlas_gene?",
         "&significant_gene?",
         "&atlas_significant_gene?",
-        "&~forbidden_gene?",
-        "&~atlas_forbidden_gene?",
+        "&~lateral_gene?",
+        "&~atlas_lateral_gene?",
     ],
     projected: Union[str, ut.Matrix] = "projected",
     reproducible: bool,
@@ -579,7 +579,7 @@ def compute_metacells_projection_correlation(
     Sets the ``projected_correlation`` per-observation annotation to the correlation between the corrected and the
     projected UMIs for each metacell. Correlation only looks at a subset of the genes specified by the
     ``mask_names``; by default, it looks only at genes common to the atlas and the query, that were "significant" in
-    both, and that were not forbidden to be used as feature in either.
+    both, and that were not lateral to be used as feature in either.
 
     If ``reproducible``, a slower (still parallel) but reproducible algorithm will be used.
     """